# Move the value assignment of vector x in gemv_n_sve.c to the outermos…